Fix issues with merge

+++ MundoBalloonApi.GraphQL/Users/UserEventMutations.AddToEventCart.cs
@@ -8,10 +8,10 @@ public partial class UserEventMutations
 {
     [Authorize]
     public Task<EventCartDetail> AddToEventCart(
-        int productVariantId, int userEventId, double quantity,
+        int productVariantId, int userEventId, decimal quantity,
         [Service] IUsersService usersService,
         CancellationToken cancellationToken)
     {
-        return usersService.AddToEventCart(productVariantId, userEventId,quantity, cancellationToken);
+        return usersService.AddToEventCart(productVariantId, userEventId, quantity, cancellationToken);
     }
 }
